+++ Service/Manager.php
@@ -641,6 +641,27 @@ class Manager
     }
 
     /**
+     * Calls "Get Settings API" in Elasticsearch and will return you the currently configured settings.
+     *
+     * return array
+     */
+    public function getSettings()
+    {
+        return $this->getClient()->indices()->getSettings(['index' => $this->getIndexName()]);
+    }
+
+    /**
+     * Gets Elasticsearch aliases information.
+     * @param $params
+     *
+     * @return array
+     */
+    public function getAliases($params = [])
+    {
+        return $this->getClient()->indices()->getAliases(array_merge(['index' => $this->getIndexName()], $params));
+    }
+
+    /**
      * Resolves type name by class name.
      *
      * @param string $className