+++ Tests/Functional/DependencyInjection/ElasticsearchExtensionTest.php
@@ -55,17 +55,4 @@ class ElasticsearchExtensionTest extends WebTestCase
         $this->assertTrue($container->has($id), 'Container should have set id.');
         $this->assertInstanceOf($instance, $container->get($id), 'Container has wrong instance set to id.');
     }
-
-    /**
-     * Test if container sets the default values as expected.
-     */
-    public function testContainerDefaultParams()
-    {
-        $container = $this->createClient()->getContainer();
-
-        $expectedManagers = ['default'];
-        $actualManagers = $container->getParameter('es.managers');
-
-        $this->assertEquals($expectedManagers, array_keys($actualManagers));
-    }
 }